// frontend/clients/generated/types/assetMeta.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import { PublicKey } from '@metaplex-foundation/umi';
import {
  Serializer,
  bool,
  bytes,
  publicKey as publicKeySerializer,
  struct,
  u16,
  u8,
} from '@metaplex-foundation/umi/serializers';

export type AssetMeta = {
  mint: PublicKey;
  pythPrice: PublicKey;
  pythFeedId: Uint8Array;
  decimals: number;
  enabledAsCollateral: boolean;
  index: number;
};

export type AssetMetaArgs = AssetMeta;

export function getAssetMetaSerializer(): Serializer<AssetMetaArgs, AssetMeta> {
  return struct<AssetMeta>(
    [
      ['mint', publicKeySerializer()],
      ['pythPrice', publicKeySerializer()],
      ['pythFeedId', bytes({ size: 66 })],
      ['decimals', u8()],
      ['enabledAsCollateral', bool()],
      ['index', u16()],
    ],
    { description: 'AssetMeta' }
  ) as Serializer<AssetMetaArgs, AssetMeta>;
}